+++ src/Ext/TechnoType/Body.h
@@ -176,6 +176,7 @@ public:
 	Nullable<int> ShadowSizeCharacteristicHeight;
 
 	Valueable<CSFText> EnemyUIName;
+	Valueable<TechnoTypeClass*> FakeOf;
 
 	bool ForceWeapon_Check;
 	Valueable<int> ForceWeapon_Naval_Decloaked;
@@ -572,6 +573,7 @@ public:
 		, CombatAlert_EVA {}
 
 		, EnemyUIName {}
+		, FakeOf {}
 
 		, VoiceCreated {}