+++ src/Phobos.h
@@ -119,6 +119,20 @@ public:
 		static bool ShowGameTime;
 		static int ShowGameTime_BoardOpacity;
 		static bool SelectCapturedCommand;
+		static bool AllowSwitchNoMoveCommand;
+		static bool AllowDistributionCommand;
+		static bool AllowDistributionCommand_SpreadModeHotKey;
+		static bool AllowDistributionCommand_SpreadModeScroll;
+		static bool AllowDistributionCommand_SpreadModeDrag;
+		static bool AllowDistributionCommand_UseClick;
+		static bool AllowDistributionCommand_FilterMode;
+		static bool AllowDistributionCommand_AffectsAllies;
+		static bool AllowDistributionCommand_AffectsEnemies;
+		static bool AllowDistributionCommand_AffectsNeutral;
+		static bool ApplyNoMoveCommand;
+		static unsigned int DistributionSpreadRange;
+		static unsigned int DistributionSpreadScrollStep;
+		static int DistributionFilterMode;
 	};
 
 	class Misc